Promote metrics as a top level command
