+++ lib/spoom/cli.rb
@@ -6,6 +6,7 @@ require "thor"
 require_relative "cli/commands/bump"
 require_relative "cli/commands/config"
 require_relative "cli/commands/lsp"
+require_relative "cli/commands/metrics"
 require_relative "cli/commands/run"
 
 module Spoom
@@ -25,6 +26,9 @@ module Spoom
 
       desc "lsp", "send LSP requests to Sorbet"
       subcommand "lsp", Spoom::Cli::Commands::LSP
+
+      desc "metrics", "collect metrics related to Sorbet"
+      subcommand "metrics", Spoom::Cli::Commands::Metrics
 
       desc "tc", "run Sorbet and parses its output"
       subcommand "tc", Spoom::Cli::Commands::Run
